Modèle Representation +Relation ManyToOne with Location & Show

+++ src/main/java/be/iccbxl/pid/repository/ShowRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import be.iccbxl.pid.models.Location;
 import be.iccbxl.pid.models.Show;
 
-public interface ShowRepository extends CrudRepository<Show, Long> {
+public interface ShowRepository extends CrudRepository<Show,Integer> {
 	Show findBySlug(String slug);
 	Show findByTitle(String title);
 	List<Show> findByLocation(Location location);
